requirements.txt
altgraph==0.17.4
bleach==6.2.0
bokeh==3.7.3
bottle==0.13.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.2
idna==3.10
Jinja2==3.1.6
jsoneditor==1.6.0
linkify-it-py==2.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.40.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
panel==1.7.0
param==2.2.0
pefile==2023.2.7
pillow==11.2.1
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pyviz_comms==3.0.4
pywebview==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
setuptools==80.8.0
six==1.17.0
tornado==6.5
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
webencodings==0.5.1
xyzservices==2025.4.0
